@@ -1,5 +1,5 @@
 {
-    "name": "proai/eloquent-versioning",
+    "name": "jpnut/eloquent-versioning",
     "description": "An extension for the Eloquent ORM to support versioning.",
     "keywords": [
         "laravel",
@@ -12,37 +12,33 @@
         "audit",
         "audits"
     ],
-    "homepage": "http://github.com/ProAI/eloquent-versioning",
+    "homepage": "http://github.com/jpnut/eloquent-versioning",
     "license": "MIT",
     "authors": [
         {
             "name": "Markus J. Wetzel",
             "email": "markuswetzel@gmx.net"
+        }, {
+            "name": "Jamie Nuttall",
+            "email": "nuttalljamie@gmail.com"
         }
     ],
     "require": {
-        "php": ">=7.1",
-        "illuminate/database": "5.*"
+        "php": "^7.4",
+        "laravel/framework": "^6.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.0",
-        "mockery/mockery": "^1.0",
-        "orchestra/testbench": "^3.6",
-        "orchestra/database": "^3.6"
+        "orchestra/testbench": "^4.0",
+        "phpunit/phpunit": "^8.5"
     },
     "autoload": {
         "psr-4": {
-            "ProAI\\Versioning\\": "src/"
+            "JPNut\\Versioning\\": "src/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "ProAI\\Versioning\\Tests\\": "tests/"
-        }
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "1.0-dev"
+            "JPNut\\Versioning\\Tests\\": "tests/"
         }
     }
 }
