@@ -1,7 +1,7 @@
 <?php
 
 use Faker\Generator as Faker;
-use ProAI\Versioning\Tests\Models\User;
+use JPNut\Versioning\Tests\Models\Dummy;
 
 /*
 |--------------------------------------------------------------------------
@@ -9,10 +9,10 @@ use ProAI\Versioning\Tests\Models\User;
 |--------------------------------------------------------------------------
 |
 */
-$factory->define(User::class, function (Faker $faker) {
+$factory->define(Dummy::class, function (Faker $faker) {
     return [
-        'email'         => $faker->unique()->safeEmail,
-        'username'      => $faker->userName,
-        'city'          => $faker->city
+        'email' => $faker->unique()->safeEmail,
+        'name'  => $faker->userName,
+        'city'  => $faker->city
     ];
 });
